+++ relay-server/src/services/store.rs
@@ -97,11 +97,15 @@ pub struct StoreMetrics {
     pub mode: ExtractionMode,
 }
 
+#[derive(Debug)]
+pub struct StoreCogs(pub Vec<u8>);
+
 /// Service interface for the [`StoreEnvelope`] message.
 #[derive(Debug)]
 pub enum Store {
     Envelope(StoreEnvelope),
     Metrics(StoreMetrics),
+    Cogs(StoreCogs),
 }
 
 impl Interface for Store {}
@@ -119,6 +123,14 @@ impl FromMessage<StoreMetrics> for Store {
 
     fn from_message(message: StoreMetrics, _: ()) -> Self {
         Self::Metrics(message)
+    }
+}
+
+impl FromMessage<StoreCogs> for Store {
+    type Response = NoResponse;
+
+    fn from_message(message: StoreCogs, _: ()) -> Self {
+        Self::Cogs(message)
     }
 }
 
@@ -149,6 +161,7 @@ impl StoreService {
         match message {
             Store::Envelope(message) => self.handle_store_envelope(message),
             Store::Metrics(message) => self.handle_store_metrics(message),
+            Store::Cogs(message) => self.handle_store_cogs(message),
         }
     }
 
@@ -396,6 +409,16 @@ impl StoreService {
                 dropped,
                 scoping,
                 Outcome::Invalid(DiscardReason::Internal),
+            );
+        }
+    }
+
+    fn handle_store_cogs(&self, StoreCogs(payload): StoreCogs) {
+        let message = KafkaMessage::Cogs(CogsKafkaMessage(payload));
+        if let Err(error) = self.produce(KafkaTopic::Cogs, 0, message) {
+            relay_log::error!(
+                error = &error as &dyn std::error::Error,
+                "failed to store cogs measurement"
             );
         }
     }
@@ -1598,6 +1621,10 @@ struct MetricsSummaryKafkaMessage<'a> {
     tags: &'a BTreeMap<String, String>,
 }
 
+#[derive(Debug, Serialize)]
+#[serde(transparent)]
+struct CogsKafkaMessage(Vec<u8>);
+
 /// An enum over all possible ingest messages.
 #[derive(Debug, Serialize)]
 #[serde(tag = "type", rename_all = "snake_case")]
@@ -1620,6 +1647,7 @@ enum KafkaMessage<'a> {
     CheckIn(CheckInKafkaMessage),
     Span(SpanKafkaMessage<'a>),
     MetricsSummary(MetricsSummaryKafkaMessage<'a>),
+    Cogs(CogsKafkaMessage),
 }
 
 impl Message for KafkaMessage<'_> {
@@ -1637,6 +1665,7 @@ impl Message for KafkaMessage<'_> {
             KafkaMessage::CheckIn(_) => "check_in",
             KafkaMessage::Span(_) => "span",
             KafkaMessage::MetricsSummary(_) => "metrics_summary",
+            KafkaMessage::Cogs(_) => "cogs",
         }
     }
 
@@ -1660,7 +1689,8 @@ impl Message for KafkaMessage<'_> {
             | Self::Profile(_)
             | Self::ReplayRecordingNotChunked(_)
             | Self::Span(_)
-            | Self::MetricsSummary(_) => Uuid::nil(),
+            | Self::MetricsSummary(_)
+            | Self::Cogs(_) => Uuid::nil(),
 
             // TODO(ja): Determine a partitioning key
             Self::Metric { .. } => Uuid::nil(),
@@ -1692,25 +1722,29 @@ impl Message for KafkaMessage<'_> {
     }
 
     /// Serializes the message into its binary format.
-    fn serialize(&self) -> Result<Vec<u8>, ClientError> {
+    fn serialize(&self) -> Result<Cow<'_, [u8]>, ClientError> {
         match self {
-            KafkaMessage::Session(message) => {
-                serde_json::to_vec(message).map_err(ClientError::InvalidJson)
-            }
-            KafkaMessage::Metric { message, .. } => {
-                serde_json::to_vec(message).map_err(ClientError::InvalidJson)
-            }
-            KafkaMessage::ReplayEvent(message) => {
-                serde_json::to_vec(message).map_err(ClientError::InvalidJson)
-            }
-            KafkaMessage::Span(message) => {
-                serde_json::to_vec(message).map_err(ClientError::InvalidJson)
-            }
-            KafkaMessage::MetricsSummary(message) => {
-                serde_json::to_vec(message).map_err(ClientError::InvalidJson)
-            }
+            KafkaMessage::Session(message) => serde_json::to_vec(message)
+                .map(Cow::Owned)
+                .map_err(ClientError::InvalidJson),
+            KafkaMessage::Metric { message, .. } => serde_json::to_vec(message)
+                .map(Cow::Owned)
+                .map_err(ClientError::InvalidJson),
+            KafkaMessage::ReplayEvent(message) => serde_json::to_vec(message)
+                .map(Cow::Owned)
+                .map_err(ClientError::InvalidJson),
+            KafkaMessage::Span(message) => serde_json::to_vec(message)
+                .map(Cow::Owned)
+                .map_err(ClientError::InvalidJson),
+            KafkaMessage::MetricsSummary(message) => serde_json::to_vec(message)
+                .map(Cow::Owned)
+                .map_err(ClientError::InvalidJson),
 
-            _ => rmp_serde::to_vec_named(&self).map_err(ClientError::InvalidMsgPack),
+            KafkaMessage::Cogs(CogsKafkaMessage(payload)) => Ok(payload.into()),
+
+            _ => rmp_serde::to_vec_named(&self)
+                .map(Cow::Owned)
+                .map_err(ClientError::InvalidMsgPack),
         }
     }
 }